+++ src/SWP/Bundle/CoreBundle/Model/Article.php
@@ -19,10 +19,11 @@ namespace SWP\Bundle\CoreBundle\Model;
 use SWP\Bundle\ContentBundle\Model\Article as BaseArticle;
 use SWP\Component\MultiTenancy\Model\OrganizationAwareTrait;
 use SWP\Component\MultiTenancy\Model\TenantAwareTrait;
+use SWP\Component\Paywall\Model\PaywallSecuredTrait;
 
 class Article extends BaseArticle implements ArticleInterface
 {
-    use TenantAwareTrait, OrganizationAwareTrait;
+    use TenantAwareTrait, OrganizationAwareTrait, PaywallSecuredTrait;
 
     /**
      * @var PackageInterface
@@ -115,5 +116,14 @@ class Article extends BaseArticle implements ArticleInterface
     public function setExternalArticle(ExternalArticleInterface $externalArticle): void
     {
         $this->externalArticle = $externalArticle;
+    }
+
+    public function getPackageExternalData()
+    {
+        if (null === $this->getPackage()->getExternalData()) {
+            return [];
+        }
+
+        return $this->getPackage()->getExternalData()->getData();
     }
 }