+++ src/SWP/Component/Bridge/Model/ExternalDataInterface.php
@@ -1,0 +1,22 @@
+<?php
+
+/*
+ * This file is part of the Superdesk Web Publisher Bridge Component.
+ *
+ * Copyright 2018 Sourcefabric z.ú. and contributors.
+ *
+ * For the full copyright and license information, please see the
+ * AUTHORS and LICENSE files distributed with this source code.
+ *
+ * @copyright 2018 Sourcefabric z.ú
+ * @license http://www.superdesk.org/license
+ */
+
+namespace SWP\Component\Bridge\Model;
+
+interface ExternalDataInterface
+{
+    public function getData(): array;
+
+    public function setData(array $data): void;
+}